refactor: split invariant reporter out of errors.ts into shared/invariants.ts
