+++ packages/zambdas/src/shared/index.ts
@@ -13,6 +13,7 @@ export * from './fhir';
 export * from './getAuth0Token';
 export * from './helpers';
 export * from './insurances';
+export * from './invariants';
 export * from './lambda';
 export * from './patients';
 export * from './pdf';
