+++ packages/zambdas/vitest.setup.ts
@@ -5,5 +5,9 @@ vi.mock('@sentry/aws-serverless', () => ({
   init: vi.fn(),
   isInitialized: vi.fn(() => false),
   setTag: vi.fn(),
+  setTags: vi.fn(),
+  captureException: vi.fn(),
+  captureMessage: vi.fn(),
+  withScope: vi.fn((cb: (scope: { setTag: (k: string, v: unknown) => void }) => void) => cb({ setTag: vi.fn() })),
   wrapHandler: vi.fn((handler) => handler), // Pass through the handler without modification
 }));
